+++ composeApp/src/androidMain/kotlin/com/altmann/choresmanager/ui/screens/calendar/weeklycalendar/WeeklyCalendar.android.kt
@@ -1,0 +1,8 @@
+package com.altmann.choresmanager.ui.screens.calendar.weeklycalendar
+
+import androidx.compose.foundation.ScrollState
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun ScrollBar(scrollState: ScrollState) {
+}