+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/utils/CalendarHelper.kt
@@ -4,6 +4,7 @@ import com.altmann.choresmanager.DateProvider
 import kotlinx.datetime.DatePeriod
 import kotlinx.datetime.DayOfWeek
 import kotlinx.datetime.LocalDate
+import kotlinx.datetime.LocalTime
 import kotlinx.datetime.Month
 import kotlinx.datetime.minus
 import kotlinx.datetime.plus